+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `#currency`, `#good_till_date`, `#order_type` and `#time_in_force` attributes to
   `IGMarkets::Streaming::WorkingOrderUpdate`
 - Fixed `IGMarkets::AccountState` not updating its working orders when a level update is received
+- Added `IGMarkets::Error::GetSessionTimeoutError` and `IGMarkets::Error::InvalidShareOrderInstrumentDataError`
 
 ### 0.27 — January 18, 2017